+++ lib/core/flutly_app_bar/flutly_app_bar.dart
@@ -1,11 +1,18 @@
-import 'package:flutly/core/flutly_tab_bar/flutly_tab_controller.dart';
-import 'package:flutly/core/flutly_tab_bar/flutly_tab_view_controller.dart';
-import 'package:flutly/models/flutly_variable.dart';
+import 'package:flutly/core/flutly_app_bar/flutly_app_bar_item.dart';
 import 'package:flutter/material.dart';
+import 'package:get/get.dart';
 
 class FlutlyAppBar {
   Color? color;
   double? blur;
 
-  FlutlyAppBar({this.color, this.blur});
+  List<FlutlyAppBarItem>? bars = List.empty(growable: true);
+
+  FlutlyAppBar({this.color, this.blur, this.bars});
+
+  FlutlyAppBarItem? getBar(String name){
+    if(bars == null) return null;
+    print("TARANAN APP BAR : " + name);
+    return bars!.firstWhereOrNull((element) => element.name == name);
+  }
 }