+++ lib/core/flutly_app_bar/flutly_app_bar_item.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class FlutlyAppBarItem {
+  String name;
+  Widget appBar;
+  double? appBarHeight;
+  bool? appBarAnimated;
+
+  FlutlyAppBarItem({
+    required this.name,
+    required this.appBar,
+    this.appBarHeight,
+    this.appBarAnimated,
+  });
+
+  bool isAnimatedAppBar() => appBarAnimated ?? false;
+}