+++ lib/core/flutly_tab_bar/flutly_tab_view_controller.dart
@@ -10,6 +10,6 @@ class FlutlyTabViewController {
     this.initalIndex,
     this.animationDuration,
   }) {
-    if (this.initalIndex == null) initalIndex = 0;
+    initalIndex ??= 0;
   }
 }