+++ CHANGELOG.md
@@ -1,3 +1,11 @@
-## [1.0.0+1] - 2024-09-12
+## [1.0.0+1] - 09.23.2024
 ### Added
 - Initial pre-release
+
+## [1.1.0+31] - 10.01.2024
+### Added
+- Flutly Profile Image
+- Flutly Dialog
+- Flutly Toast
+- Flutly Tab Bar Controller System
+- More Utilities